Changes naming of ned to netting-entity and moves utility
