+++ netting-entity/benchmarkNed.js
@@ -1,11 +1,9 @@
 const express = require("express");
 const cors = require("cors");
 const commander = require("commander");
-const web3Utils = require("web3-utils");
 const shell = require("shelljs");
 const fs = require('fs');
-const Utility = require("../utility-js/Utility");
-const hhHandler = require("./household-handler");
+const Utility = require("./utility");
 const zkHandler = require("./zk-handler");
 const web3Helper = require("../helpers/web3");
 const { ZERO_ADDRESS } = require("../helpers/constants");
