+++ netting-entity/household-handler.js
