+++ netting-entity/index.js
@@ -4,7 +4,7 @@ const commander = require("commander");
 const web3Utils = require("web3-utils");
 const shell = require("shelljs");
 const fs = require("fs");
-const Utility = require("../utility-js/Utility");
+const Utility = require("./utility");
 const hhHandler = require("./household-handler");
 const zkHandler = require("./zk-handler");
 const web3Helper = require("../helpers/web3");
