+++ netting-entity/utility.js
