+++ package.json
@@ -11,7 +11,7 @@
     "test-contracts:win32": ".\\test-contracts.bat",
     "run-server": "node ./household-server/index.js",
     "run-sensor": "node ./mock-sensor/index.js",
-    "run-ned": "node ./ned-server/index.js",
+    "run-netting-entity": "node ./netting-entity/index.js",
     "run-benchmark-ned": "node ./ned-server/benchmarkNed.js",
     "lint-all": "eslint *.js household-server/ mock-sensor/ test/ helpers/ household-ui/src utility-js/ scripts/ migrations/ && yarn lint-contracts",
     "lint-contracts": "solium -d contracts/",
