--- utility-js/README.md
@@ -1,6 +1,0 @@
-# Off-chain utility implementation
-
-The off-chain utility is a JavaScript implementation of the utility contracts of milestone one which is used to track energy production and consumption of households as well as compute netting results off-chain.
-\texttt{Utility.js} preserves the same functionality for tracking and netting.
-
-Run `yarn install` followed by `yarn test-utility-js` to test the functionality of the off-chain utility.
